Invert logic for selecting SnapshotHandler
Currently the private `determineHandler()` selects the `SnapshotHandler` based on the `paparazzi.test.record` system property, which the Gradle tasks set. Unfortunately this forces users to use the `verifyPaparazzi` task to perform snapshot verification. Using `testDebug` or even `check` will trigger a recording to happen instead, which feels incorrect.

This PR flips that logic around so that it checks for the `paparazzi.test.record` property instead. This results in `testDebug` performing a verification, and users needing to use the `recordPaparazzi` task to perform the record. The `verifyPaparazzi` task could probably be removed after this PR but I'll leave that for now.