Introduce method to allow modifying Paparazzi config.
The call to `Paparazzi.inflate` was using resources based on the DeviceConfig passed into Paparazzi's constructor and the resources will not be updated until a new SessionParams is built
with the updated DeviceConfig.

This change moves several parameters from the snapshot and gif methods over to a new method `Paparazzi.unsafeConfig` which ensures that a new session is created with the updated config.
If users wish to test layout resource files for different device sizes they will be required to inflate a second time after updating Paparazzi's config.

Ex:
```
  @Test
  fun mix() {
    var launch = paparazzi.inflate<LinearLayout>(R.layout.launch)
    paparazzi.snapshot(launch, "pixel3")
    paparazzi.unsafeUpdateConfig(deviceConfig = NEXUS_7)
    launch = paparazzi.inflate(R.layout.launch)
    paparazzi.snapshot(launch, "nexus7")
  }
}
```

Resolves #194