Fix incompatibility when used on compileSdk 33 projects
Caused by the SDK 33 ro.build.version.known_codenames build prop being
too long for the pre-SDK 33 SystemProperties class. The old versions of
the SystemProperties class throw exceptions for values with lengths of
92+ chars.

Since we can't update to a version of layoutlib which contains the SDK 33
changes yet (likely in Electric Eel), this PR adds a workaround which
omits the build prop when it is longer than 91 chars. This doesn't
cause any obvious issues from my testing on the Twitter for Android app.