Move the `testName` field out of the Snapshot class.
Snapshot itself doesn’t need `testName` and having it there couples
`Paparazzi.prepare()` to JUnit. The real consumers of `testName` are the
SnapshotHandler subclasses (`HtmlReportWriter` and `SnapshotVerifier`) which use it to compute the file name. Moving `testName` to the constructor of these subclasses is not an option because it will affect Paparazzi's API, but we can inject a `FileNameProvider` which will take care of computing the file when necessary.