Change the project name when building from Intellij or Android Studio
This change set updates the `paparazzi/settings.gradle` root project name
to `paparrazzi` when the project is built from Intellij or Android
Studio. Previously when attempting to create run configurations, the IDE
would set the task to `paparazzi-libs:` in accordance with the root
project name which would fail to run because the task did not match the
directory structure. This change provides a workaround for the following
[gradle issue](https://github.com/gradle/gradle/issues/16608) by
checking a property `idea.active` which is only set when building the project
from Intellij or Android Studio. This approach is documented in this
[Stack Overflow post](https://stackoverflow.com/questions/63409895/configure-gradle-plugin-based-on-future-tasks/63411337#63411337).