Fix PaparazziPluginTest.flagDebugLinkedObjectsIsOn
expected to contain:
    Objects still linked from the DelegateManager:
but was:
    > ...
    > Task :testDebugUnitTest UP-TO-DATE

    BUILD SUCCESSFUL in 424ms
    18 actionable tasks: 18 up-to-date

	at app//app.cash.paparazzi.gradle.PaparazziPluginTest.flagDebugLinkedObjectsIsOn(PaparazziPluginTest.kt:170)

This is because:
 * on the first execution it creates the build folder with intermediate files
 * on the second execution the build folder exists so everything is up to date

Solution: start with a clean slate which will force executing the test task every time.

Also apply the same fix to the similar "off" test, because it was a false positive doesNotContain on second execution: if the task is up to date, then there's no output, therefore doesNotContain wrongly passed.