Silence big warning on IDE import
```
WARNING: We recommend using a newer Android Gradle plugin to use compileSdkPreview = "UpsideDownCakePrivacySandbox"

This Android Gradle plugin (8.0.2) was tested up to compileSdk = 33.

You are strongly encouraged to update your project to use a newer
(stable or preview) Android Gradle plugin that has been tested with compileSdkPreview = "UpsideDownCakePrivacySandbox".

If you are already using the latest preview version of the Android Gradle plugin,
you may need to wait until a newer version with support for compileSdkPreview = "UpsideDownCakePrivacySandbox" is available.

To suppress this warning, add/update
    android.suppressUnsupportedCompileSdk=UpsideDownCakePrivacySandbox
to this project's gradle.properties.
```