Fully advance the choreographer on a frame
Compose UI 1.8 moved ripples into native when running with layoutlib.
The native layer is unable to see the manual clock ticks that are
currently being done and throws an exception due to the frame time
being 0. Invoking doFrame on the choreographer propagates the time to
the native layer.

This commit also cleans up clock management logic that was creating
invalid uptimes.

https://android.googlesource.com/platform/frameworks/support/+/09bea4c5dea7cf1aacf1a49b2f15595d9c159e2d